gcp-service/functions/file-aggregator/requirements.txt
functions-framework==3.*
google-cloud-storage
pymysql==1.1.0